Reorganize repo and add contributing docs in prep for making repo public
