+++ .gitignore
@@ -10,4 +10,4 @@ coverage.txt
 vendor/*/
 
 # Binaries from builds
-bin/
+build/
