+++ experimental/bootstrap/cmd/infrakitctl/container/Dockerfile
