+++ experimental/bootstrap/destroy.go
