--- plugin/instance/container/manifest.json
@@ -1,14 +1,0 @@
-{
-    "manifestVersion": "v0.1",
-    "description": "infrakit instance driver for aws",
-    "documentation": "https://docs.docker.com/engine/extend/plugins/",
-    "entrypoint": ["driver", "--listen=/run/docker/plugins/aws-instance.sock"],
-    "network": {
-        "type": "host"
-    },
-    "interface" : {
-        "types": ["docker.instanceDriver/1.0"],
-        "socket": "aws-instance.sock"
-    },
-    "mounts":[]
-}
