--- plugin/instance/container/plugin-config.json
@@ -1,6 +1,0 @@
-{
-    "mounts":[],
-    "env":[],
-    "args":null,
-    "devices":null
-}
