--- plugin/instance/container/plugins.json
@@ -1,62 +1,0 @@
-{
-    "@PLUGIN_ID@": {
-        "plugin": {
-            "Manifest": {
-                "Args": {
-                    "Value": null,
-                    "Settable": null,
-                    "Description": "",
-                    "Name": ""
-                },
-                "Env": null,
-                "Devices": null,
-                "Mounts": [],
-                "Capabilities": null,
-                "ManifestVersion": "v0.1",
-                "Description": "infrakit instance driver for aws",
-                "Documentation": "https://docs.docker.com/engine/extend/plugins/",
-                "Interface": {
-                    "Socket": "aws-instance.sock",
-                    "Types": [
-                        "docker.instanceDriver/1.0"
-                    ]
-                },
-                "Entrypoint": [ "driver", "--listen=/run/docker/plugins/aws-instance.sock" ],
-                "Workdir": "",
-                "User": {},
-                "Network": {
-                    "Type": "host"
-                }
-            },
-            "Config": {
-                "Devices": null,
-                "Args": null,
-                "Env": [],
-                "Mounts": [
-                    {
-                        "Options": null,
-                        "Type": "bind",
-                        "Destination": "/var/run",
-                        "Source": "/var/run",
-                        "Settable": null,
-                        "Description": "docker socket",
-                        "Name": ""
-                    },
-                    {
-                        "Options": null,
-                        "Type": "bind",
-                        "Destination": "/run/docker",
-                        "Source": "/run/docker",
-                        "Settable": null,
-                        "Description": "plugins socket",
-                        "Name": ""
-                    }
-                ]
-            },
-            "Active": true,
-            "Tag": "latest",
-            "Name": "@PLUGIN_REPO_OWNER@/@PLUGIN_REPO@",
-            "Id": "@PLUGIN_ID@"
-        }
-    }
-}
