+++ plugin/instance/instance_test.go
@@ -1,4 +1,4 @@
-package aws
+package instance
 
 import (
 	"encoding/json"
